refact : features/interview-record/video-record 로직 분리
